+++ Pipfile
@@ -10,6 +10,7 @@ flask = "*"
 flask-basicauth = "*"
 influxdb = "*"
 waitress = "*"
+python-dateutil = "*"
 
 [requires]
 python_version = "3.7"
